requirements.txt
asgiref==3.7.2
attr==0.3.2
brotli==1.1.0
ConfigParser==6.0.0
cryptography==3.4.8
Cython==3.0.7
dl==0.1.0
docutils==0.20.1
ecdsa==0.18.0
flaskr==1.0.0
HTMLParser==0.0.2
importlib_metadata==4.6.4
ipython==8.19.0
ipywidgets==8.1.1
jnius==1.1.0
keyring==23.5.0
lockfile==0.12.2
lxml==4.9.4
numpy==1.26.2
ordereddict==1.1
Pillow==10.1.0
protobuf==3.12.4
pycryptodome==3.19.1
pyOpenSSL==23.3.0
python-dotenv==1.0.0
railroad==0.5.0
Sphinx==7.2.6
thread==0.1.3
toml==0.10.2
tornado==6.4
watchdog==3.0.0
xmlrpclib==1.0.1
zipp==1.0.0
